adds blob storage
